python/pyproject.toml
[tool.poetry]
name = "nox"
version = "0.1.0"
description = "an attempt at a JARVIS written in Python"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.1"
requests = "^2.26.0"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
